Load Disclaimer class with Composer's autoloader
Note: I am not autoloading Options yet.
The line
require 'src/admin/Options.php';

loads the Options class so when we run

$options = new Options()

PHP already knows about the class Options, so the autoloader is never
called.

For new Disclaimer(), PHP doesn't know about the Disclaimer class and
would therefore throw an error. Instead of throwing an error, the
autoloader uses the Namespace + classname to look for the file to
autoload "CHB\DraftPostDisclaimer\Disclaimer.php"

In composer.json, we've now told the autoloader to replace
"CHB\DraftPostDisclaimer\" with "src/" so it looks for the file at
"src/Disclaimer.php", which is where we have the file defined.

See #4