Use autoloader to load Options class
Options class with namespace was
"CHB\DraftPostDisclaimer\Options"
based on our "autoload" value in composer.json, the autoloader looks for
this at
"src/Options.php"
however that is NOT the path to the file.

In order to get the autoloader to work, we need to fix this disconnect.

We could have fixed it by moving Options out of the "/admin" directory
and up to
"src/Options.php"
but I chose to take a different approach.

First, I modified the namespace in the Options.php to match the
filepath.

Second, I changed the instantiation of Options in
draft-post-disclaimer.php because the namespace is
different than the namespace of Options. I could have changed the
instantation to use the full namespace
$options = new \CHB\DraftPostDisclaimer\Admin\Options();
but instead I went with the version relative to the current namespace
$options = new Admin\Options();

See #4